requirements.txt
# Django and database
Django==4.2.9
djangorestframework==3.14.0
django-import-export==3.3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0

# LangChain and LLM tools
langchain==0.0.352
openai==1.30.1
tiktoken==0.5.2
faiss-cpu==1.7.4

# PDF handling
PyPDF2==3.0.1

# Environment variables
python-dotenv==1.0.1

# Plotly charts (optional for chart features)
pandas==2.2.2
plotly==5.22.0

# For xhtml2pdf reports (optional)
xhtml2pdf==0.2.12
